added task "1296. Divide Array in Sets of K Consecutive Numbers"

+++ Algorithms/0846.HandOfStraights/solution.cpp
@@ -22,7 +22,7 @@ public:
             int prevCard = iterator->first;
             --iterator->second;
             iterator = checkMoveNext(cardMap, iterator);
-            for (size_t groupIndex = 1; groupIndex < groupSize; ++groupIndex)
+            for (size_t groupIndex = 1; groupIndex < static_cast<size_t>(groupSize); ++groupIndex)
             {
                 if (iterator == cardMap.end())
                     return false;
